+++ addon/utils/make-bound-helper.js
@@ -1,9 +1,9 @@
 import Ember from 'ember';
 
 export default function(helper) {
-	if (Ember.Helper && Ember.Helper.detect(helper)) {
-		return helper;
-	}
+  if (Ember.Helper && Ember.Helper.detect(helper)) {
+    return helper;
+  }
 
-	return Ember.HTMLBars.makeBoundHelper(helper);
+  return Ember.HTMLBars.makeBoundHelper(helper);
 }