+++ app/helpers/moment-from-now.js
@@ -1,0 +1,4 @@
+import momentFromNowHelper from 'ember-moment/helpers/moment-from-now';
+import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
+
+export default makeBoundHelper(momentFromNowHelper);