+++ config/ember-try.js
@@ -7,14 +7,14 @@ module.exports = {
         'ember-load-initializers': 'ember-cli/ember-load-initializers#0.0.2'
       }
     },
-		{
+    {
       name: 'ember-1.11.3',
       dependencies: {
         'ember': '1.11.3',
         'ember-load-initializers': 'ember-cli/ember-load-initializers#0.0.2'
       }
     },
-		{
+    {
       name: 'ember-1.12.0',
       dependencies: {
         'ember': '1.12.0'