+++ tests/dummy/app/controllers/index.js
@@ -1,15 +1,15 @@
 import Ember from 'ember';
-import computedDuration from 'ember-moment/computeds/duration';
-import computedMoment from 'ember-moment/computeds/moment';
-import computedAgo from 'ember-moment/computeds/ago';
+import momentDuration from 'ember-moment/computeds/duration';
+import momentFormat from 'ember-moment/computeds/format';
+import momentFromNow from 'ember-moment/computeds/from-now';
 
 export default Ember.Controller.extend({
   now: new Date(),
   lastHour: new Date(new Date().valueOf() - (60*60*1000)),
   date: new Date(),
   numHours: 822,
-  computedDate: computedMoment('date', 'MM/DD/YY hh:mm:ss'),
-  computedOneHourAgo: computedAgo('lastHour'),
-  computedNumHours: computedDuration('numHours', 'hours'),
+  computedDate: momentFormat('date', 'MM/DD/YY hh:mm:ss'),
+  computedOneHourAgo: momentFromNow('lastHour'),
+  computedNumHours: momentDuration('numHours', 'hours'),
   usIndependenceDay: new Date(1776, 6, 4, 12, 0, 0)
 });